refine code & add readme

+++ Controllers/DeptCotroller.js
@@ -29,13 +29,21 @@ const deptController = {
     create: async (req, res) => {
         try {
             const { name } = req.body;
+
+            // Validate empty fields
+            if (!name) {
+                return res.status(400).json({ message: 'All fields are required' });
+            }
+            
             const is_exist = await Department.findOne({ name })
             if (is_exist) {
-                return res.status(200).json({ message: 'Department already exist' })
+                return res.status(409).json({ message: 'Department already exist' })
             }
-            const newDept = await Department({ name });
+
+            let formattedDept = name.charAt(0).toUpperCase() + name.slice(1)
+            const newDept = await Department({ name:formattedDept });
             newDept.save();
-            return res.status(200).json({ message: 'Department created successfully', "Department": newDept })
+            return res.status(201).json({ message: 'Department created successfully', "Department": newDept })
         } catch (error) {
             console.log('Error creating department', error);
             return res.status(500).json({ message: 'Internal server error' })
@@ -46,17 +54,19 @@ const deptController = {
             const { name } = req.body;
             const validateDept = await Department.findOne({_id: req.params.id});
 
-            if (validateDept.name === 'newcomer') {
-                return res.status(400).json({ message: 'Default department can not be updated' })
+            if (validateDept.name === 'Newcomer') {
+                return res.status(403).json({ message: 'Default department can not be updated' })
             }
 
-            const updateDept = await Department.findByIdAndUpdate(req.params.id, {name});
+            let formattedDept = name.charAt(0).toUpperCase() + name.slice(1)
+
+            const updateDept = await Department.findByIdAndUpdate(req.params.id, {name:formattedDept});
 
             if (!updateDept) {
                 return res.status(400).json({ message: 'Department not found' })
             }
 
-            return res.json({ message: 'Department updated successfully', department: updateDept });
+            return res.status(200).json({ message: 'Department updated successfully', department: updateDept });
         } catch (error) {
             console.log('Error updating department', error);
             return res.json({ message: 'Internal server error' });
@@ -64,6 +74,13 @@ const deptController = {
     },
     delete: async (req, res) => {
         try {
+            const validateDept = await Department.findOne({_id: req.params.id});
+            const validateUser = await User.findOne({username: req.user.username}).populate('department', 'name');
+
+            if (validateUser.department.name === validateDept.name){
+                return res.status(403).json({ message: "You can not delete associate department." })
+            }
+
             const deleteDept = await Department.findByIdAndDelete(req.params.id)
 
             if (!deleteDept) {
@@ -72,7 +89,7 @@ const deptController = {
 
             await User.deleteMany({department:deleteDept._id})
 
-            return res.json({ message: 'Department and associated employees deleted successfully', department: deleteDept });
+            return res.status(200).json({ message: 'Department and associated employees deleted successfully', department: deleteDept });
 
         } catch (error) {
             console.log('Error deleting department', error);
